refactor(#17): reorganize project structure
- Move menus to 'controller' package
- Move services to 'service' package
- Move items to 'model' package
- Update imports accordingly
- Delete `Item` from model names
