+++ .gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,2 @@
-#Tue Nov 25 22:06:57 NOVT 2025
-gradle.version=8.14.3
+#Sun Nov 30 19:56:29 NOVT 2025
+gradle.version=8.10
