--- src/main/java/app/Main.java
@@ -1,48 +1,0 @@
-package app;
-
-import java.util.Scanner;
-
-import schedule.ScheduleMenu;
-import schedule.ScheduleService;
-import tasks.TaskService;
-import tasks.TaskMenu;
-import pomodoro.PomodoroService;
-import pomodoro.PomodoroMenu;
-
-public class Main{
-    public static void main(String[] args){
-        Scanner scanner=new Scanner(System.in);
-
-        ScheduleService scheduleService=new ScheduleService();
-        ScheduleMenu scheduleMenu=new ScheduleMenu(scheduleService,scanner);
-
-        TaskService taskService = new TaskService();
-        TaskMenu taskMenu =new TaskMenu(taskService, scanner);
-
-        PomodoroService pomodoroService = new PomodoroService();
-        PomodoroMenu pomodoroMenu = new PomodoroMenu(pomodoroService, taskService, scanner);
-        
-        while(true){
-            System.out.println("\n=== Главное меню ===");
-            System.out.println("1. Расписание");
-            System.out.println("2. Задачи");
-            System.out.println("3. Pomodoro");
-            System.out.println("0. Выход");
-            System.out.print("Выбор: ");
-            String choice=scanner.nextLine().trim();
-
-            switch(choice){
-                case "1" -> scheduleMenu.menu();
-                case "2" -> taskMenu.menu();
-                case "3" -> pomodoroMenu.menu();
-                case "0" -> {
-                    System.out.println("Выход...");
-                    scanner.close();
-                    return;
-                }
-                
-                default -> System.out.println("Неверный выбор, попробуйте снова.");
-            }
-        }
-    }
-}
